post

+++ app/Http/Controllers/TasksController.php
@@ -16,12 +16,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //
-         $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -52,12 +60,16 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        // $task = new Task;
+        // $task->content = $request->content;
+        // $task->status = $request->status;
+        // $task->save();
+        $request->user()->tasks()->create([
+            'content'=>$request->content,
+            'status'=>$request->status,
+            ]);
 
-        return redirect('/');
+        return redirect()->back();
     }
 
     /**
@@ -124,8 +136,11 @@ class TasksController extends Controller
     {
         //
         $task = task::find($id);
-        $task->delete();
+        // $task->delete();
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        }
 
-        return redirect('/');
+        return redirect()->back();
     }
 }
